- Added following user with camera animation

+++ app/src/main/java/com/example/googlemapscompose/presentation/mapscreen/MapState.kt
@@ -10,6 +10,6 @@ data class MapState(
 )
 
 data class LocationState(
-    val latitude: Float = 0.0f,
-    val longitude: Float = 0.0f,
+    val latitude: Double = 0.0,
+    val longitude: Double = 0.0
 )
